requirements.txt
streamlit~=1.45.0
pandas~=2.2.3
numpy~=2.2.5
pytz~=2025.2
altair~=5.5.0
matplotlib~=3.10.1
yfinance~=0.2.59
quantstats-lumi~=0.3.3
scipy~=1.15.2
scikit-learn~=1.6.1
arch==7.2.0
joblib~=1.5.0
ipython~=9.2.0
pandas-ta==0.3.14b
ta~=0.11.0
beautifulsoup4~=4.13.4
soupsieve~=2.7
packaging~=24.2
typing_extensions~=4.13.2
charset-normalizer~=3.4.2
GitPython~=3.1.44
gitdb~=4.0.12
pillow~=11.2.1
pip~=21.3.1
wheel~=0.37.1
protobuf~=5.29.4
tornado~=6.4.2
cffi~=1.17.1
setuptools~=60.2.0
parso~=0.8.4
jedi~=0.19.2
colorama~=0.4.6
toml~=0.10.2
attrs~=25.3.0
smmap~=5.0.2
threadpoolctl~=3.6.0
platformdirs~=4.3.8
jsonschema~=4.23.0
narwhals~=1.38.1
pyarrow~=20.0.0
Jinja2~=3.1.6
referencing~=0.36.2
traitlets~=5.14.3
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
certifi~=2025.4.26
Pygments~=2.19.1
ipython_pygments_lexers~=1.1.1
stack-data~=0.6.3
decorator~=5.2.1
prompt_toolkit~=3.0.51
wcwidth~=0.2.13
peewee~=3.18.1
tzdata~=2025.2
seaborn~=0.13.2
cycler~=0.12.1
urllib3~=2.4.0
six~=1.17.0
idna~=3.10
tenacity~=9.1.2
watchdog~=6.0.0
curl_cffi~=0.10.0
frozendict~=2.4.6
websockets~=15.0.1
multitasking~=0.0.11
requests~=2.32.3
asttokens~=3.0.0
contourpy~=1.3.2
executing~=2.2.0
fonttools~=4.57.0
pure_eval~=0.2.3
pycparser~=2.22
pyparsing~=3.2.3
click~=8.1.8
cachetools~=5.5.2
blinker~=1.9.0
jsonschema-specifications~=2025.4.1
kiwisolver~=1.4.8
tabulate~=0.9.0